fix for rendering and no info
