+++ client/src/components/friends-page/FriendsContent.css
@@ -81,7 +81,7 @@
 }
 
 .no-mates-message {
-  margin: 100px auto;
+  margin-top: 100px;
   font-family: var(--cookie-text-title);
   font-size: 50px;
   text-align: center;
