+++ client/src/components/single-dog/SingleDog.css
@@ -50,7 +50,7 @@
 .options-wrapper {
   display: flex;
   justify-self: end;
-  margin-top: 60%;
+  margin-top: 30%;
 }
 
 .single-dog-options {
